Cleanup model upload
