+++ manager/src/main/scala/io/hydrosphere/serving/manager/service/model/ModelManagementServiceImpl.scala
@@ -1,23 +1,20 @@
 package io.hydrosphere.serving.manager.service.model
 
-import java.nio.file.{Files, Path, Paths}
 import java.time.LocalDateTime
 
 import io.hydrosphere.serving.contract.model_contract.ModelContract
 import io.hydrosphere.serving.contract.utils.description.ContractDescription
 import io.hydrosphere.serving.contract.utils.ops.ModelContractOps._
-import io.hydrosphere.serving.manager.controller.model.{ModelUpload, UploadedEntity}
-import io.hydrosphere.serving.manager.model.Result.ClientError
+import io.hydrosphere.serving.manager.controller.model.ModelUpload
 import io.hydrosphere.serving.manager.model._
 import io.hydrosphere.serving.manager.model.api.{ModelMetadata, ModelType}
 import io.hydrosphere.serving.manager.model.db.Model
 import io.hydrosphere.serving.manager.repository._
 import io.hydrosphere.serving.manager.service.contract.ContractUtilityService
 import io.hydrosphere.serving.manager.service.source.fetchers.ModelFetcher
-import io.hydrosphere.serving.manager.service.source.sources.ModelSource
-import io.hydrosphere.serving.manager.service.source.SourceManagementService
-import io.hydrosphere.serving.manager.util.TarGzUtils
-import Result.Implicits._
+import io.hydrosphere.serving.manager.service.source.ModelStorageManagementService
+import cats.data.EitherT
+import cats.implicits._
 import org.apache.logging.log4j.scala.Logging
 import spray.json.JsObject
 
@@ -28,7 +25,7 @@ import scala.util.{Failure, Success}
 class ModelManagementServiceImpl(
   modelRepository: ModelRepository,
   modelVersionRepository: ModelVersionRepository,
-  sourceManagementService: SourceManagementService,
+  sourceManagementService: ModelStorageManagementService,
   contractService: ContractUtilityService
 )(
   implicit val ex: ExecutionContext
@@ -140,78 +137,26 @@ class ModelManagementServiceImpl(
   }
 
   override def modelContractDescription(modelId: Long): HFResult[ContractDescription] = {
-    getMap(getModel(modelId)) { model =>
-      model.modelContract.flatten
-    }
+    val f = for {
+      model <- EitherT(getModel(modelId))
+    } yield model.modelContract.flatten
+    f.value
   }
 
-  def writeFilesToSource(source: ModelSource, files: Map[Path, Path]): Unit = {
-    files.foreach {
-      case (src, dest) =>
-        source.writeFile(dest.toString, src.toFile)
-    }
-  }
-
-  def uploadToSource(upload: ModelUpload): HFResult[CreateOrUpdateModelRequest] = {
-    val fMaybeSource = upload.source match {
-      case Some(sourceName) => sourceManagementService.getSource(sourceName)
-      case None => sourceManagementService.getSources.map(_.headOption.toHResult(ClientError("No sources available")))
-    }
-    fMaybeSource.map { result =>
-      result.right.map { source =>
-        val uploadName = upload.tarballPath.getFileName.toString
-        val unpackDir = Files.createTempDirectory(uploadName)
-        val rootDir = Paths.get(uploadName)
-        val uploadedFiles = TarGzUtils.decompress(upload.tarballPath, unpackDir)
-        val localFiles = uploadedFiles
-          .filter(_.startsWith(unpackDir))
-          .map { path =>
-            val relPath = unpackDir.relativize(path)
-            path -> rootDir.resolve(relPath)
-          }
-          .toMap
-
-        writeFilesToSource(source, localFiles)
-
-        val inferredMeta = ModelFetcher.fetch(source, unpackDir.toString)
-        val contract = upload.contract.getOrElse(inferredMeta.contract)
-        val modelType = upload.modelType.map(ModelType.fromTag).getOrElse(inferredMeta.modelType)
-        val modelName = upload.name.getOrElse(inferredMeta.modelName)
-
-        CreateOrUpdateModelRequest(
-          id = None,
-          name = modelName,
-          source = source.sourceDef.name,
-          modelType = modelType,
-          description = upload.description,
-          modelContract = contract
-        )
-      }
-    }
-  }
-
-  override def uploadModelTarball(upload: ModelUpload): HFResult[Model] = {
-    uploadToSource(upload).flatMap {
-      case Right(request) =>
-        modelRepository.get(request.name).flatMap {
-          case Some(model) =>
-            val updateRequest = request.copy(id = Some(model.id), source = model.source)
-            logger.info(s"Updating uploaded model with id: ${updateRequest.id} name: ${updateRequest.name}, source: ${updateRequest.source}, type: ${updateRequest.modelType} ")
-            updateModelRequest(updateRequest)
-          case None =>
-            val newSource = s"${request.source}:${upload.name}"
-            val createRequest = request.copy(source = newSource)
-            logger.info(s"Creating uploaded model with name: ${createRequest.name}, source: ${createRequest.source}, type: ${createRequest.modelType}")
-            createModel(createRequest)
-        }
-      case Left(err) => Future.successful(Left(err))
-    }
-  }
-
-  private def getMap[T, R](generator: => HFResult[T])(callback: T => R): HFResult[R] = {
-    generator.map { result =>
-      result.right.map(callback)
-    }
+  override def uploadModel(upload: ModelUpload): HFResult[Model] = {
+    val f = for {
+      result <- EitherT(sourceManagementService.upload(upload))
+      request = CreateOrUpdateModelRequest(
+        id = None,
+        name = result.name,
+        source = result.source,
+        modelType = result.modelType,
+        description = result.description,
+        modelContract = result.modelContract
+      )
+      r <- EitherT(upsertRequest(request))
+    } yield r
+    f.value
   }
 
   override def indexModels(ids: Set[Long]): Future[Seq[IndexStatus]] = {
@@ -274,6 +219,20 @@ class ModelManagementServiceImpl(
               modelRepository.delete(model.id).map(_ => ModelDeleted(model))
           }
       }
+    }
+  }
+
+  private def upsertRequest(request: CreateOrUpdateModelRequest): Future[Either[Result.HError, Model]] = {
+    modelRepository.get(request.name).flatMap {
+      case Some(model) =>
+        val updateRequest = request.copy(id = Some(model.id), source = model.source)
+        logger.info(s"Updating uploaded model with id: ${updateRequest.id} name: ${updateRequest.name}, source: ${updateRequest.source}, type: ${updateRequest.modelType}")
+        updateModelRequest(updateRequest)
+      case None =>
+        val newSource = s"${request.source}:${request.name}"
+        val createRequest = request.copy(source = newSource)
+        logger.info(s"Creating uploaded model with name: ${createRequest.name}, source: ${createRequest.source}, type: ${createRequest.modelType}")
+        createModel(createRequest)
     }
   }
 }
