+++ manager/src/main/scala/io/hydrosphere/serving/manager/service/source/SourceUploadResult.scala
@@ -1,0 +1,12 @@
+package io.hydrosphere.serving.manager.service.source
+
+import io.hydrosphere.serving.contract.model_contract.ModelContract
+import io.hydrosphere.serving.manager.model.api.ModelType
+
+case class SourceUploadResult(
+  name: String,
+  source: String,
+  modelType: ModelType,
+  description: Option[String],
+  modelContract: ModelContract
+)
